requirements.txt
alpaca-py>=0.38.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
apscheduler>=3.10.4
httpx>=0.28.0
python-telegram-bot>=21.9
python-dotenv>=1.0.1
pytz>=2024.2
pytest>=8.3.0
pytest-asyncio>=0.24.0
